+++ beliefs/models/base_models.py
@@ -1,8 +1,41 @@
-import copy
 import networkx as nx
 
-from beliefs.models.DirectedGraph import DirectedGraph
 from beliefs.utils.math_helper import is_kronecker_delta
+
+
+class DirectedGraph(nx.DiGraph):
+    """
+    Base class for all directed graphical models.
+    """
+    def __init__(self, edges=None, node_labels=None):
+        """
+        Input:
+            edges: an edge list, e.g. [(parent1, child1), (parent1, child2)]
+            node_labels: a list of strings of node labels
+        """
+        super().__init__()
+        if edges is not None:
+            self.add_edges_from(edges)
+        if node_labels is not None:
+            self.add_nodes_from(node_labels)
+
+    def get_leaves(self):
+        """
+        Returns a list of leaves of the graph.
+        """
+        return [node for node, out_degree in self.out_degree() if out_degree == 0]
+
+    def get_roots(self):
+        """
+        Returns a list of roots of the graph.
+        """
+        return [node for node, in_degree in self.in_degree() if in_degree == 0]
+
+    def get_topologically_sorted_nodes(self, reverse=False):
+        if reverse:
+            return list(reversed(list(nx.topological_sort(self))))
+        else:
+            return nx.topological_sort(self)
 
 
 class BayesianModel(DirectedGraph):
@@ -69,8 +102,8 @@ class BayesianModel(DirectedGraph):
         return vars_in_definite_state - observed
 
     def _get_ancestors_of(self, observed):
-        """Return list of ancestors of observed labels, including the observed labels themselves."""
-        ancestors = observed.copy()
+        """Return list of ancestors of observed labels"""
+        ancestors = set()
         for label in observed:
             ancestors.update(nx.ancestors(self, label))
         return ancestors
@@ -87,7 +120,9 @@ class BayesianModel(DirectedGraph):
           reachable_observed_vars: set of strings, observed labels (variables with direct
               evidence) that are reachable from the source label.
         """
+        # ancestors of observed labels, including observed labels
         ancestors_of_observed = self._get_ancestors_of(observed)
+        ancestors_of_observed.update(observed)
 
         visit_list = set()
         visit_list.add((source, 'up'))
